pyproject.toml
[project]
name = "prime-trade"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "python-binance>=1.0.29",
]
